feature: separate each class for main_window and editor_window
